Summary
Added You method that sets values for firstName, lastName, favoriteAnimal, and age. I created the method getFullName that returns the first and last name that is saved with a space in between. I also created setfavoriteAnimal that sets the input string to the private variable favoriteAnimal. Lastly, I made a happyBirthday method that prints "Happy Birthday!" along with adding to age by 1 and returning the age.